requirements.txt
torch>=1.7.0
torchvision>=0.8.1
ultralytics>=8.0.0
opencv-python-headless>=4.5.0
numpy>=1.21.0
Pillow>=8.0.0
PyYAML>=5.3
requests>=2.23.0
scipy>=1.4.1
matplotlib>=3.2.2
seaborn>=0.11.0
pandas>=1.1.4
pyzmq>=25.0.0